Half-hearted attempt to try the Minion font.
The OTF files threw an exception on load (loca table not found), and
when converted to TTF (using a website) they didn't render properly
(the characters were the wrong ones). This is probably a bug in
PdfBox, since I just pass them the font they loaded.

Reverting back to Times New Roman, but keeping the Typeface enum.